A major update. A lot of functionality ironed out, no tests for this. Basic best find exists as a function test, can also be used as a template for a GUI program later, one scenario.

+++ src/main/java/se/good_omens/EliteDangerous_TraderHelper/common/interfaces/Commodity.java
@@ -18,7 +18,7 @@ public interface Commodity extends XmlEntity {
 	public Boolean isIllegal();
 	public Boolean isRare();
 
-	public long getAveragePrice();
-	public long getSellingPrice();
-	public long getBuyingPrice();
+	public int getAveragePrice();
+	public int getSellingPrice();
+	public int getBuyingPrice();
 }
